requirements.txt
Flask
pandas 
praw
nltk
requests
psycopg2-binary 
apache-airflow-providers-postgres
sqlalchemy
alpaca-py